# digital-laundry-backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "Laundry Management API", "version": "1.0.0" },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/register/student": {
      "post": {
        "summary": "Register student",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "rollNumber",
                  "email",
                  "hostelCode",
                  "roomNo",
                  "password"
                ],
                "properties": {
                  "fullName": { "type": "string" },
                  "rollNumber": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "hostelCode": { "type": "string" },
                  "roomNo": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registered. Verify email." },
          "400": { "description": "Invalid" }
        }
      }
    },
    "/api/auth/register/staff": {
      "post": {
        "summary": "Register staff",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fullName", "email", "staffCode", "password"],
                "properties": {
                  "fullName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "staffCode": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registered. Verify email." },
          "400": { "description": "Invalid" }
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "summary": "Verify email OTP",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "otp": { "type": "string", "minLength": 6, "maxLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email verified" },
          "400": { "description": "Invalid/expired OTP" }
        }
      }
    },
    "/api/auth/resend-otp": {
      "post": {
        "summary": "Resend OTP",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT token" },
          "400": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Forgot password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Temporary password sent" } }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "Change password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["oldPassword", "newPassword"],
                "properties": {
                  "oldPassword": { "type": "string" },
                  "newPassword": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password changed" },
          "400": { "description": "Old password incorrect" }
        }
      }
    },
...
    "/api/complaints/{orderId}": {
      "post": {
        "summary": "Create complaint for order (PENDING only)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["description"],
                "properties": { "description": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Complaint created" },
          "400": { "description": "Invalid" }
        }
      }
    },
    "/api/complaints": {
      "get": {
        "summary": "List complaints",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/complaints/{id}": {
      "get": {
        "summary": "Get complaint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Complaint with images" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/complaints/{id}/images": {
      "post": {
        "summary": "Upload complaint images (JSON array of base64/URLs)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["images"],
                "properties": {
                  "images": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Uploaded" } }
      }
    },
    "/api/complaints/{id}/advance": {
      "post": {
        "summary": "Advance complaint (OPEN→IN_REVIEW→RESOLVED)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Advanced" },
          "400": { "description": "Invalid transition" }
        }
      }
    },
    "/api/complaints/{id}/close": {
      "post": {
        "summary": "Close complaint (RESOLVED→CLOSED)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Closed" },
          "400": { "description": "Only RESOLVED → CLOSED" }
        }
      }
    },
...
  }
}